+++ engines/scumm/scumm.h
@@ -955,6 +955,8 @@ protected:
 	void drawStripToScreen(VirtScreen *vs, int x, int width, int top, int bottom);
 	void mac_drawStripToScreen(VirtScreen *vs, int top, int x, int y, int width, int height);
 	void mac_restoreCharsetBg();
+	void mac_drawLoomPracticeMode();
+
 	void ditherCGA(byte *dst, int dstPitch, int x, int y, int width, int height) const;
 
 public:
